Add base types to gem
