+++ apps/api/app/graphql/mutations/base_mutation.rb
@@ -1,8 +1,8 @@
 module Mutations
   class BaseMutation < Souls::SoulsMutation
-    argument_class Types::BaseArgument
-    field_class Types::BaseField
-    input_object_class Types::BaseInputObject
-    object_class Types::BaseObject
+    argument_class Souls::Types::BaseArgument
+    field_class Souls::Types::BaseField
+    input_object_class Souls::Types::BaseInputObject
+    object_class Souls::Types::BaseObject
   end
 end
