--- apps/api/app/graphql/types/base/base_input_object.rb
@@ -1,5 +1,0 @@
-module Types
-  class BaseInputObject < GraphQL::Schema::InputObject
-    argument_class Types::BaseArgument
-  end
-end
