--- apps/api/app/graphql/types/base/base_interface.rb
@@ -1,7 +1,0 @@
-module Types
-  module BaseInterface
-    include GraphQL::Schema::Interface
-
-    field_class Types::BaseField
-  end
-end
