--- apps/api/app/graphql/types/base/base_object.rb
@@ -1,6 +1,0 @@
-module Types
-  class BaseObject < GraphQL::Schema::Object
-    field_class Types::BaseField
-    connection_type_class Types::BaseConnection
-  end
-end
