+++ apps/api/app/graphql/types/base/mutation_type.rb
@@ -1,5 +1,5 @@
 module Types
-  class MutationType < Types::BaseObject
+  class MutationType < Souls::Types::BaseObject
     SoulsHelper.get_tables.each do |t|
       %w[create update delete destroy_delete].each do |a|
         field "#{a}_#{t.singularize.underscore}".to_sym,
