+++ apps/api/app/graphql/types/connections/base_connection.rb
@@ -1,14 +1,4 @@
 module Types
-  class BaseConnection < GraphQL::Types::Relay::BaseConnection
-    field :total_count, Integer, null: false
-    field :total_pages, Integer, null: false
-
-    def total_count
-      object.items.size
-    end
-
-    def total_pages
-      (total_count / object.max_page_size) + 1
-    end
+  class BaseConnection < Souls::SoulsConnection
   end
 end
