--- apps/worker/app/graphql/types/base/base_argument.rb
@@ -1,4 +1,0 @@
-module Types
-  class BaseArgument < GraphQL::Schema::Argument
-  end
-end
