--- apps/worker/app/graphql/types/base/base_enum.rb
@@ -1,4 +1,0 @@
-module Types
-  class BaseEnum < GraphQL::Schema::Enum
-  end
-end
