--- apps/worker/app/graphql/types/base/base_field.rb
@@ -1,5 +1,0 @@
-module Types
-  class BaseField < GraphQL::Schema::Field
-    argument_class Types::BaseArgument
-  end
-end
