--- apps/worker/app/graphql/types/base/base_object.rb
@@ -1,5 +1,0 @@
-module Types
-  class BaseObject < GraphQL::Schema::Object
-    field_class Types::BaseField
-  end
-end
