--- apps/worker/app/graphql/types/base/base_scalar.rb
@@ -1,4 +1,0 @@
-module Types
-  class BaseScalar < GraphQL::Schema::Scalar
-  end
-end
