--- apps/worker/app/graphql/types/base/base_union.rb
@@ -1,4 +1,0 @@
-module Types
-  class BaseUnion < GraphQL::Schema::Union
-  end
-end
