+++ apps/worker/app/graphql/types/base/query_type.rb
@@ -1,5 +1,5 @@
 module Types
-  class QueryType < Types::BaseObject
+  class QueryType < Souls::Types::BaseObject
     add_field(GraphQL::Types::Relay::NodeField)
     add_field(GraphQL::Types::Relay::NodesField)
     workers =
