+++ lib/souls/app/graphql/types/base_argument.rb
@@ -1,0 +1,6 @@
+module Souls
+  module Types
+    class BaseArgument < GraphQL::Schema::Argument
+    end
+  end
+end
