+++ lib/souls/app/graphql/types/base_enum.rb
@@ -1,0 +1,6 @@
+module Souls
+  module Types
+    class BaseEnum < GraphQL::Schema::Enum
+    end
+  end
+end
