+++ lib/souls/app/graphql/types/base_field.rb
@@ -1,0 +1,7 @@
+module Souls
+  module Types
+    class BaseField < GraphQL::Schema::Field
+      argument_class Souls::Types::BaseArgument
+    end
+  end
+end
