+++ lib/souls/app/graphql/types/base_input_object.rb
@@ -1,0 +1,7 @@
+module Souls
+  module Types
+    class BaseInputObject < GraphQL::Schema::InputObject
+      argument_class Souls::Types::BaseArgument
+    end
+  end
+end
