+++ lib/souls/app/graphql/types/base_interface.rb
@@ -1,0 +1,9 @@
+module Souls
+  module Types
+    module BaseInterface
+      include GraphQL::Schema::Interface
+
+      field_class Souls::Types::BaseField
+    end
+  end
+end
