+++ lib/souls/app/graphql/types/base_scalar.rb
@@ -1,0 +1,6 @@
+module Souls
+  module Types
+    class BaseScalar < GraphQL::Schema::Scalar
+    end
+  end
+end
