+++ lib/souls/app/graphql/types/base_union.rb
@@ -1,0 +1,6 @@
+module Souls
+  module Types
+    class BaseUnion < GraphQL::Schema::Union
+    end
+  end
+end
