+++ lib/souls/app/graphql/types/index.rb
@@ -1,0 +1,11 @@
+require_relative "base_argument"
+require_relative "base_enum"
+require_relative "base_field"
+require_relative "base_input_object"
+require_relative "base_interface"
+require_relative "base_object"
+require_relative "base_scalar"
+require_relative "base_union"
+
+module Souls
+end
