+++ lib/souls/app/index.rb
@@ -1,7 +1,9 @@
 require_relative "graphql/souls_mutation"
 require_relative "graphql/souls_query"
+require_relative "graphql/souls_connection"
 require_relative "utils/souls_logger"
 require_relative "utils/firebase_id_token"
+require_relative "graphql/types/index"
 
 module Souls
 end
